hal: sdk-ng: sync sdk usb to hal nxp

+++ mcux/mcux-sdk-ng/middleware/usb/host/class/usb_host_cdc_ecm.c
@@ -114,7 +114,7 @@ usb_status_t USB_HostCdcEcmDataSend(usb_host_class_handle classHandle,
 }
 
 usb_status_t USB_HostCdcEcmGetMacStringDescriptor(usb_host_class_handle classHandle,
-                                                  uint16_t index,
+                                                  uint8_t index,
                                                   uint16_t langID,
                                                   uint8_t *buffer,
                                                   uint16_t bufferLength,
